Refactor bin startup scripts to remove duplication
