+++ CHANGELOG.md
@@ -1,0 +1,15 @@
+# Change Log
+
+All notable changes to this project will be documented in this file according to [Keep A Changelog](http://keepachangelog.com/).
+
+This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+
+### Added
+
+- CHANGELOG.md
+
+### Changed
+
+- [In progress] Refactored bin/env.php and bin/utils.php
